Stop testing IE9.
We are seeing issues on SauceLabs where `stop` seems to be invalidating after a
few hundred ms despite having a 30s timeout.  When it happens, it seems to
happen exactly once: the whole suite passes except for one test with a timeout
error but timings that are off by orders of magnitude.

This seems to happen only in IE9.