Custom inline adapters can change wrapping.
This is useful for Conductor.js so it can swap out the global Oasis.

We could do so in Oasis generally, but this likely requires more thought.  The
general problem is we don't have dynamic scoping, so we can't really deal with
the problem properly (swapping out the global temporarily does nothing for
asynchronous connection, for instance).

At the moment, the inline adapter works, but is not easy to use correctly.