Drop polyglot.
Previously Oasis sandboxes could be defined entirely in JavaScript, even when
loaded in the iframe adapter.  This was convenient for certain potential
Conductor.js cases but led to a fair bit of complexity.

The type of a sandbox is now implied by the adapter that is used to load it.
For the iframe_adapter this means all sandboxes now are of type html, so the
user is responsible for loading oasis.

This also means that there is no longer any sandbox bootstrapping, so for the
web worker adapter, users must now load Oasis, their own dependencies and their
sandbox.  To load a previous web worker sandbox `wwsandbox.js` something like
the following is now required:

```js
importScripts('oasis.js');
importScripts('wwsandbox.js');
```

see https://gist.github.com/hjdivad/45fa4be9f5aaf56586ea