Reject handlers with no consumer when connecting before receiving ports.
When Oasis and the card are both loaded as a file in the sandbox, it is
possible that the card tries to connect the consumers before the ports are
actually received. In that case, we still need to reject capabilities
that have no corresponding consumers.

With the polyglot approach, ports were sent with the card's url in the
initialization message. Therefore the capabilities in the sandbox were
registered before the card actually loads and connects the ports.